requirements.txt
Flask==2.3.2
pillow==10.1.0
pytesseract==0.3.10

